feat: fixed soundsystem, fixed emergency switches for 0.10.x hardware

+++ Firmware/LowLevel/include/soundsystem.h
@@ -40,7 +40,7 @@ class MP3Sound
 
     public:
 
-                uint32_t anzSoundfiles;          // number of files stored on the SD-card
+                int16_t anzSoundfiles;          // number of files stored on the SD-card
                 bool     playing;
 
                 MP3Sound();
